[JIT] python IR bindings: consolidate tests, add short docs in OVERVIEW.md
Document the existence of python IR bindings; quick comments about it; and consolidate tests in one file to serve as examples to users.

[ghstack-poisoned]
